% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donneesHuguesBenoit.R
\name{donneesHuguesBenoit}
\alias{donneesHuguesBenoit}
\title{Produit un fichier données utilisées par Hugues Benoit pour ses travaux sur la mortalité naturelle et survie suite au rejet par la pêche}
\usage{
donneesHuguesBenoit(sourceRecap, destination)
}
\arguments{
\item{destination}{chemin d'accès où enregistrée les données formattées}

\item{source}{chemin d'accès de la base de données originale}
}
\value{
le dataframe enregistré
}
\description{
Produit un fichier données utilisées par Hugues Benoit pour ses travaux sur la mortalité naturelle et survie suite au rejet par la pêche
}
\examples{
print('Non disponible')
}
