% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_cph.R
\name{lecture_cph}
\alias{lecture_cph}
\title{Lire la base de données des conditions d'hamecons remontés. Effectue une uniformisation des abréviations des espèces vers le francais.
Les abbréviations attendues sont a=appât, b=bait, c=cod, e=empty, f=flétan, h=halibut, i=???, m=morue, r=redfish, s=sébastes, t=turbot, v=vide, x=autres.
À FAIRE: ajouter les entrées pour 3Pn
À FAIRE: valider les "stations supplémentaires", soit des stations de pêche (non scientifique) avec couverture d'observateur similiare au relevé
(durant premières années)
À FAIRE: valider les entrées 'i'}
\usage{
lecture_cph(origine, destination)
}
\arguments{
\item{origine}{localisation et nom du fichier de base de données des stations échantillonnées}

\item{destination}{localisation et nom du fichier de travail local}
}
\value{
une list()
}
\description{
Lire la base de données des conditions d'hamecons remontés. Effectue une uniformisation des abréviations des espèces vers le francais.
Les abbréviations attendues sont a=appât, b=bait, c=cod, e=empty, f=flétan, h=halibut, i=???, m=morue, r=redfish, s=sébastes, t=turbot, v=vide, x=autres.
À FAIRE: ajouter les entrées pour 3Pn
À FAIRE: valider les "stations supplémentaires", soit des stations de pêche (non scientifique) avec couverture d'observateur similiare au relevé
(durant premières années)
À FAIRE: valider les entrées 'i'
}
