% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_longueurs.R
\name{lecture_longueurs}
\alias{lecture_longueurs}
\title{Lecture de la base de données de longueur du relevé à la palangre.}
\usage{
lecture_longueurs(
  origine,
  lpFletan_alpha = 5.262e-06,
  lpFletan_beta = 3.167,
  lpMorue_alpha = 5.74e-06,
  lpMorue_beta = 3.116,
  donneesSupplementaires = FALSE
)
}
\arguments{
\item{origine}{chemin d'accès de la base de données}

\item{lpFletan_alpha}{`numeric` alpha de la relation longueur-poids du flétan de l'Atlantique: `poids=alpha*longueur^beta`}

\item{lpFletan_beta}{`numeric` beta de la relation longueur-poids du flétan de l'Atlantique: `poids=alpha*longueur^beta`}

\item{lpMorue_alpha}{`numeric` alpha de la relation longueur-poids de la morue: `poids=alpha*longueur^beta`}

\item{lpMorue_beta}{`numeric` beta de la relation longueur-poids de la morue: `poids=alpha*longueur^beta`}

\item{donneesSupplementaires}{Si les données recueillies dans les stations où le poisson est conservé doivent être inclues}
}
\value{
`data.frame` de la base de données lue et mise en forme
}
\description{
Contient toutes les espèces mesurées durant le relevé.
Des données supplémentaires ont aussi récoltées durant la pêche scientifique où des flétan sont conservés pour financement.
}
\details{
À faire: considérer 3Pn
}
\examples{
print('non disponible')
}
