% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_stations.R
\name{lecture_stations}
\alias{lecture_stations}
\title{Lit la base de données des stations}
\usage{
lecture_stations(origine, destination)
}
\arguments{
\item{origine}{chemin d'accès de la base de données}

\item{destination}{chemin d'accès où enregistrer la base de données de travail}
}
\value{
`data.frame` mis en forme
}
\description{
Lit la base de données des stations
}
\examples{
print('non disponible')
}
