% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_versionsTravail.R
\name{lecture_versionTravail}
\alias{lecture_versionTravail}
\title{Lectures des bases de données et création d'une version de travail (locale) des données.}
\usage{
lecture_versionTravail(annee, dir_bd = NULL)
}
\arguments{
\item{annee}{integer de la dernière année a considérer}

\item{dir_bd}{position du dossier contenant la base de donnée}
}
\value{
une liste des stations, longueurs et cph
}
\description{
Les données de stations, de longueur, et de captures par hamecons sont lues et organisées en objets utilisables dans les autres fonctions du package.
Ces objets sont enregistrés en format .csv dans le répertoire fourni
}
