% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_carte_protocole.R
\name{produire_carte_protocole}
\alias{produire_carte_protocole}
\title{Production de la carte à joindre au protocole du relevé à la palangre}
\usage{
produire_carte_protocole(annee, dir_releve, dir_pbs, dir_shapefile)
}
\arguments{
\item{annee}{anne pour laquelle la carte est calculée}

\item{dir_releve}{chemin pour le dossier ou se trouve les protocoles, pour lire et écrire}

\item{dir_pbs}{chemin pour le dossier ou se trouve les outils pré-formatés en format PBSmapping}
}
\description{
La fonction lit les données de cartographie, les stations.
La fonction enregistre un tableau des stations pour inclure dans le protocole.
La fonction enregistre des cartes pour inclure dans le protocole, soit une générale et plusieurs aggrandissements.
}
\examples{
## à venir
}
