% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirage_stations.R
\name{tirage_stations}
\alias{tirage_stations}
\title{Tirage aléatoire des stations du relevé à la palangre}
\usage{
tirage_stations(
  annee,
  dir_releve,
  dir_shapefile,
  fichier_nb_stations,
  fichier_raster,
  seed = NULL
)
}
\arguments{
\item{annee}{année pour laquelle les stations sont tirées.}

\item{dir_releve}{chemin du répertoire où les stations tirées seront enregistrées}

\item{dir_shapefile}{chemin du répertoire où lire les shapefiles de la zone d'étude}

\item{fichier_nb_stations}{chemin du fichier .csv contenant le nombre de stations allouées à chaque strate}

\item{fichier_raster}{chemin du fichier contenant les informations de profondeur}

\item{seed}{valeur utilisée pour initialiser le tirage aléatoire des stations. Si NULL, le paramètre `annee` est utilisé.#'}
}
\value{
un data.frame des stations tirées
}
\description{
Tirage aléatoire des stations du relevé à la palangre
}
\examples{
#à venir
}
