% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_longueurs.R
\name{lecture_longueurs}
\alias{lecture_longueurs}
\title{Lecture de la base de données de longueur du relevé à la palangre.}
\usage{
lecture_longueurs(
  origine,
  destination,
  lpFletan_alpha = 5.262e-06,
  lpFletan_beta = 3.167,
  lpMorue_alpha = 5.74e-06,
  lpMorue_beta = 3.116,
  donneesSupplementaires = FALSE
)
}
\arguments{
\item{donneesSupplementaires}{Si les données recueillies dans les stations où le poisson est conservé doivent être inclues}
}
\description{
Contient toutes les espèces mesurées durant le relevé.
Des données supplémentaires ont aussi récoltées durant la pêche scientifique où des flétan sont conservés pour financement.
}
\details{
À faire: considérer 3Pn
}
